add a further simplifying wrapper: useGetSinglePreviewImage
